Create sample chapter
This includes the content for the sample chapter:

* Introduction
* How DNS works
* Dig
* Transferring a domain
* Remove www from a domain
* Use a CNAME on the Apex domain

Links to "When to use CNAMES", "ALIAS or ANAME", "Glossary", and
"Installing Tools" appeared in these sections, so I created their
headers so the links went somewhere. Along with the headers is a short
description of what the sections contain to give readers an idea of what
was linked to.